requirements.txt
streamlit==1.31.0
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.12.0
httpx==0.27.2
google-generativeai==0.3.2
requests==2.31.0
docx2pdf==0.1.8
xhtml2pdf==0.2.15
python-dotenv==1.0.0
Jinja2==3.1.4
pytest==8.3.3